refactor: modularize bot architecture
